requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
plotly>=5.13.0
networkx>=3.0
numpy>=1.23.0
openai>=0.27.0
python-dotenv>=0.19.0
fpdf>=1.7.2
markdown>=3.4.0
matplotlib>=3.7.0 